updated cc, glibc package to prevent RPATHing
The glibc package in particular contains checks to ensure that
DT_RUNPATH and DT_RPATH are not set. This runs counter to Spack's
intended method of operation; the approach taken in this commit
should be reconsidered, but the results are good.

The cc change ensures no RPATHs are added by Spack. The package
updates ensure a) that GCC is the only compiler that can be used,
per glibc recommendations, and that the specs file it supports
creating is used to override the -rpath behavior set up by the
specs file installed by Spack.